Refactor and disable enhanced services/controllers
Renamed enhanced service and controller files to .bak and commented out their module imports to disable them. Added PartialEq to FileType enum, implemented Display for UserRole, and improved user service with get_user_by_email and create_user functions. Fixed minor issues in session, cache, and auth services for better reliability.

+++ backend/src/services/cache_service.rs
@@ -42,7 +42,7 @@ impl CacheService {
         let data = serde_json::to_string(value)
             .map_err(|e| format!("Failed to serialize value: {}", e))?;
         
-        conn.set_ex(key, data, expiration.as_secs() as usize)
+        conn.set_ex(key, data, expiration.as_secs())
             .await
             .map_err(|e| format!("Failed to set cache: {}", e))?;
